+++ src/app/form/[form_type]/[page_type]/[team_id]/page.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import PostEditor from "@/app/form/[form_type]/_components/PostEditor";
+
+const Page = () => {
+  return (
+      <PostEditor isEditMode={true}/>
+  );
+};
+
+export default Page;
