+++ src/app/form/[form_type]/[page_type]/post/page.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import PostEditor from "@/app/form/[form_type]/_components/PostEditor";
+
+const Page = () => {
+  return (
+      <PostEditor isEditMode={false}/>
+  );
+};
+
+export default Page;
