+++ src/app/form/[form_type]/_components/CreateTeamForm.tsx
@@ -1,0 +1,81 @@
+"use client";
+import {useForm, SubmitHandler, DefaultValues, Path, UseFormRegister} from "react-hook-form";
+import TextareaField from "@/components/common/Input/TextArea/TextareaField";
+import {ProjectForm, MentoringForm, ProjectFormData, MentoringFormData} from "@/app/form/[form_type]/_type/formDataTypes";
+import FormFieldRenderer from "@/app/form/[form_type]/_components/FormFieldRenderer";
+import FormRowRenderer from "@/app/form/[form_type]/_components/FormRowRenderer";
+
+interface Props<T extends ProjectFormData | MentoringFormData> {
+  onSubmit: SubmitHandler<T>;
+  defaultValues: DefaultValues<T>;
+  formFields: (ProjectForm | MentoringForm)[];
+  division?: "select" | "stacks";
+}
+
+type CustomErrors = {
+  contents?: { message: string };
+  content?: { message: string };
+};
+
+const CreateTeamForm = <T extends ProjectFormData | MentoringFormData> ({
+  onSubmit,
+  defaultValues,
+  formFields,
+  division
+}: Props<T>) => {
+  const {
+    control,
+    register,
+    handleSubmit,
+    formState: { errors },
+  } = useForm<T>({
+    defaultValues: defaultValues as unknown as DefaultValues<T>
+  });
+
+  const customErrors = errors as CustomErrors;
+
+  return (
+    <form onSubmit={handleSubmit(onSubmit)}>
+      {formFields.map((field, index) =>
+          "row" in field ? (
+              <FormRowRenderer
+                  key={index}
+                  rowFields={field.row}
+                  control={control}
+                  register={register}
+                  division={division}
+              />
+          ) : (
+              <FormFieldRenderer
+                  key={index}
+                  field={field}
+                  control={control}
+                  register={register}
+                  division={division}
+              />
+          )
+      )}
+
+      {/* 소개 */}
+      <TextareaField<ProjectFormData>
+        label="소개"
+        name={"contents" as Path<ProjectFormData>}
+        placeholder="프로젝트 소개를 입력해 주세요."
+        register={register as unknown as UseFormRegister<ProjectFormData>}
+        error={customErrors.contents?.message || customErrors.content?.message}
+      />
+
+      {/* 버튼 */}
+      <div className="mt-16 text-center md:flex">
+        <button className="w-[320px] h-[50px] mx-5 rounded-xl border-2 hover:bg-gray-400">
+          닫기
+        </button>
+        <button className="w-[320px] h-[50px] mx-5 bg-blue-500 text-white rounded-xl hover:bg-blue-400">
+          팀 생성하기
+        </button>
+      </div>
+    </form>
+  );
+};
+
+export default CreateTeamForm;
