+++ src/app/team/[page_type]/[team_id]/(external)/layout.tsx
