+++ src/app/team/[page_type]/[team_id]/(external)/viewer/_components/InfoSection.tsx
