+++ src/app/team/[page_type]/[team_id]/(member)/_type/teamPageMember.ts
