+++ src/app/team/[page_type]/[team_id]/(member)/info/_components/TeamInfoSection.tsx
