+++ src/app/team/[page_type]/[team_id]/(member)/post/_components/PostCard.tsx
